Removed ant error
Problem: ant was not building. The locale was ste to POSIX.
Solution: set the locale to en_US.utf8

Problem: The build context delays the docker build
Solution: Move the Dockerfile into its work folder so it does not copy all
          files.

Problem: The git repository is big and cloning takes a long time.
Solution: Dockerfile clones only the latest commit.