Add Gradle Wrapper
Gradle wrapper is essential for a gradle project because it helps users
who doesn't have Gradle or Ant installed to build the project just with
the JDK without the hassle of installing gradle or ant.

This helps the users who aren't Java developers and just want to use
Loklak on their machine.

Fixes https://github.com/loklak/loklak_server/issues/897