Implement Settings page
`DAO` has been modified to be able to modify configuration.

A method to save the user configuration file has been added to
`LoklakServer`, and The method has been added to Loklak's Shutdown
Hook.

`SettingsServlet` has been modified and it's used as the backend API
for modifying and viewing the current server configuration. It can
only be accessed if the server was accessed via `localhost`.

`settings.html` and `settings.js` has been added as the front-end
loklak server configuration page which communicates with
`SettingsServlet` via a simple REST API that was mentioned before.

Closes https://github.com/loklak/loklak_server/issues/876